refactor: UI 로직을 View와 Component로 분리 - lottoTicket 컴포넌트로 추출 - LottoPurchaseView로 렌더링 로직 분리 - lottoController의 로또 개수 변수 제거
