+++ src/step2-index.js
@@ -1,6 +1,9 @@
 import { LOTTO_PRICE } from "./constants/lottoInfo";
 import LottoController from "./controller/LottoController";
 import Validator from "./utils/Validator";
+import { renderLottoTicket } from "./view/web/LottoPurchaseView";
+
+const lottoController = new LottoController();
 
 const purchaseForm = document.querySelector("form");
 const purchasePriceInput = document.querySelector("#purchase-price");
@@ -16,8 +19,7 @@ const handlePurchase = (event) => {
 
     purchaseForm.reset();
     const lottoCount = purchasedPrice / LOTTO_PRICE;
-    const lottoController = new LottoController(lottoCount);
-    const purchasedLottos = lottoController.issueLottos();
+    const purchasedLottos = lottoController.issueLottos(lottoCount);
     renderLottoTicket(purchasedLottos);
   } catch (error) {
     alert(error.message);
@@ -25,24 +27,3 @@ const handlePurchase = (event) => {
 };
 
 purchaseForm.addEventListener("submit", handlePurchase);
-
-const lottoTicketContainer = document.querySelector(".lotto-ticket-container");
-
-const lottoTicket = (lotto) => {
-  const numbers = lotto.join(", ");
-
-  return `
-    <li class="lotto-ticket">
-      <span class="lotto-icon">🎟️</span>
-      <span class="lotto-numbers">${numbers}</span>
-    </li>
-  `;
-};
-
-const renderLottoTicket = (lottos) => {
-  lottoTicketContainer.innerHTML = `
-  <p>총 ${lottos.length}개를 구매하였습니다.</p>
-  <ul>
-    ${lottos.map((lotto) => lottoTicket(lotto)).join("")}
-  </ul>`;
-};
